fix(nylas): use the content disposition provided
This causes a massive perf boost since we do way less calls for files.
There is a PR in nylas-node (https://github.com/nylas/nylas-nodejs/pull/80/commits/40242a333dbe31738194b1474daad0c6a26053dd)
that fixes the issue of passing the content disposition but since we needed it quicker
I've patched and published a version myself.
We'll switch back to the official module when the PR lands.