fix(office365): catch error when failing to fetch attachments
This happens when a message is being moved while being fetched. The message id was still valid when it was received as a webhook, but at the time of fetching attachments, it was moved and the id changed.
Resulting in a 404. The error handler throwed silently, causing the unimail function callback to never be called.